Separate ArtistController from AlbumsViewerApiController and add more unit tests for AlbumsViewerApiController
