+++ src/AlbumViewerNetCore/AlbumViewerNetCore.csproj
@@ -11,6 +11,7 @@
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore.App" />
 		<PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="2.1.0" />
+		<PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.1" />
 		<PackageReference Include="Serilog.Extensions.Logging" Version="2.0.2" />
 		<PackageReference Include="Serilog.Sinks.RollingFile" Version="3.3.0" />		
 	    <ProjectReference Include="..\AlbumViewerBusiness\AlbumViewerBusiness.csproj" />
